requirements.txt
kafka-python==2.0.2
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
feast==0.32.1
faiss-cpu==1.7.4
fastapi==0.103.1
uvicorn==0.23.2
xgboost==1.7.6
pydantic==2.3.0
psycopg2-binary==2.9.7
redis==5.0.0
pyspark==3.4.1
